--- MODELS/NASR/CSV/FssDataModels.cs
@@ -1,304 +1,0 @@
-namespace FAA_DATA_HANDLER.Models.NASR.CSV
-{
-    public class FssDataModel
-    {
-        #region Common Fields
-        public class CommonFields
-        {
-            /// <summary>
-            /// Effective Date
-            /// _Src: All Apt_*.csv files(EFF_DATE)
-            /// _MaxLength: 10
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>The 28 Day NASR Subscription Effective Date in format ‘YYYY/MM/DD’.</remarks>
-            public string EffDate { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: All Fss_*.csv files(FSS_ID)
-            /// _MaxLength: 4
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string FssId { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: All Fss_*.csv files(NAME)
-            /// _MaxLength: 30
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string Name { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: All Fss_*.csv files(CITY)
-            /// _MaxLength: 40
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string City { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: All Fss_*.csv files(STATE_CODE)
-            /// _MaxLength: 2
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? StateCode { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: All Fss_*.csv files(COUNTRY_CODE)
-            /// _MaxLength: 2
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string CountryCode { get; set; }
-
-        }
-        #endregion
-
-        #region Fss_BASE Fields
-        public class FssBase : CommonFields
-        {
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(UPDATE_DATE)
-            /// _MaxLength: 10
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? UpdateDate { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(FSS_FAC_TYPE)
-            /// _MaxLength: 8
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string FssFacType { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(VOICE_CALL)
-            /// _MaxLength: 30
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? VoiceCall { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LAT_DEG)
-            /// _MaxLength: (2,0)
-            /// _DataType: int
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public int LatDeg { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LAT_MIN)
-            /// _MaxLength: (2,0)
-            /// _DataType: int
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public int LatMin { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LAT_SEC)
-            /// _MaxLength: (6,4)
-            /// _DataType: double
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public double LatSec { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LAT_HEMIS)
-            /// _MaxLength: 1
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string LatHemis { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LAT_DECIMAL)
-            /// _MaxLength: (10,8)
-            /// _DataType: double
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public double LatDecimal { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LONG_DEG)
-            /// _MaxLength: (3,0)
-            /// _DataType: int
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public int LongDeg { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LONG_MIN)
-            /// _MaxLength: (2,0)
-            /// _DataType: int
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public int LongMin { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LONG_SEC)
-            /// _MaxLength: (6,4)
-            /// _DataType: double
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public double LongSec { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LONG_HEMIS)
-            /// _MaxLength: 1
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string LongHemis { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(LONG_DECIMAL)
-            /// _MaxLength: (11,8)
-            /// _DataType: double
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public double LongDecimal { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(OPR_HOURS)
-            /// _MaxLength: 65
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string OprHours { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(FAC_STATUS)
-            /// _MaxLength: 1
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? FacStatus { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(ALTERNATE_FSS)
-            /// _MaxLength: 4
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? AlternateFss { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(WEA_RADAR_FLAG)
-            /// _MaxLength: 1
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? WeaRadarFlag { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(PHONE_NO)
-            /// _MaxLength: 16
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? PhoneNo { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_BASE.csv(TOLL_FREE_NO)
-            /// _MaxLength: 16
-            /// _DataType: string
-            /// _Nullable: Yes
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string? TollFreeNo { get; set; }
-
-        }
-        #endregion
-
-        #region Fss_RMK Fields
-        public class FssRmk : CommonFields
-        {
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_RMK.csv(REF_COL_NAME)
-            /// _MaxLength: 30
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string RefColName { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_RMK.csv(REF_COL_SEQ_NO)
-            /// _MaxLength: (3,0)
-            /// _DataType: int
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public int RefColSeqNo { get; set; }
-
-            /// <summary>
-            /// NoTitleYet
-            /// _Src: FSS_RMK.csv(REMARK)
-            /// _MaxLength: 300
-            /// _DataType: string
-            /// _Nullable: No
-            /// </summary>
-            /// <remarks>NoRemarksYet</remarks>
-            public string Remark { get; set; }
-
-        }
-        #endregion
-
-    }
-}
