+++ MODELS/NASR/CSV/LidCsvDataModels.cs
@@ -1,6 +1,6 @@
 namespace FAA_DATA_HANDLER.Models.NASR.CSV
 {
-    public class LidDataModel
+    public class LidCsvDataModel
     {
         #region Common Fields
         public class CommonFields
