+++ MODELS/NASR/CSV/RdrCsvDataModels.cs
@@ -1,6 +1,6 @@
 namespace FAA_DATA_HANDLER.Models.NASR.CSV
 {
-    public class RdrDataModel
+    public class RdrCsvDataModel
     {
         #region Common Fields
         public class CommonFields
