+++ PARSERS/NASR/CSV/AptCsvParser.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using static FAA_DATA_HANDLER.Models.NASR.CSV.AptDataModel;
+using static FAA_DATA_HANDLER.Models.NASR.CSV.AptCsvDataModel;
 
 namespace FAA_DATA_HANDLER.Parsers.NASR.CSV
 {
